adding features

+++ src/backend/libs/justin-data-starter/src/main/java/com/nttdata/pocdata/ContactsDao.java
@@ -4,7 +4,6 @@ import com.nttdata.data.Contacts;
 
 import org.springframework.data.repository.CrudRepository;
 
+import java.math.BigInteger;
 
-import java.math.BigDecimal;
-
-public interface ContactsDao extends CrudRepository<Contacts, BigDecimal> { }
+public interface ContactsDao extends CrudRepository<Contacts, BigInteger> { }
