+++ src/backend/libs/justin-data-starter/src/main/java/com/nttdata/pocdata/ContactsService.java
@@ -2,16 +2,16 @@ package com.nttdata.pocdata;
 
 import com.nttdata.data.Contacts;
 
-import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.util.List;
 import java.util.Optional;
 
 public interface ContactsService {
     List<Contacts> getContacts();
 
-    Optional<Contacts> getContact(BigDecimal contactId);
+    Optional<Contacts> getContact(BigInteger contactId);
 
     Contacts saveContact(Contacts contact);
 
-    void deleteContact(BigDecimal contactId);
+    void deleteContact(BigInteger contactId);
 }
